Do not flush() the writer.
The directive most likely passes the writer it received (the env writer) to the body, in which case there's no need to flush. Alternatively, the directive supplies another writer it owns. In that case, the directive needs to flush/close it once when it's done with it (which can be after several iterations). Automatically flushing causes performance penalty if the directive renders the nested content multiple times. Also, if we're rendering to a HttpServletResponse writer, we can cause it to commit the response to the TCP socket prematurely, which would make some people not happy :-)